+++ NEWS.md
@@ -15,11 +15,12 @@ per the updated version of dependency package 'raster'
 
 ### Test environments
 * local OS Windows, R 4.1.0
-* Fedora Linux (R-devel), R 4.1.0
+* Debian Linux, R-release, GCC, 4.1.0
+* Fedora Linux (R-devel), GCC, R 4.1.0 
 * win-builder (devel and release), R 4.1.0
 
 ### R CMD check results
-There were no ERRORs, WARNINGs or 1 NOTEs. 
+There were no ERRORs, WARNINGs or 1 NOTES. 
 
 **NOTEs 1**: The note pertains to using a lot of non-default packages. 
 ***Author's comment***: This will be addressed in