+++ NEWS.md
@@ -2,6 +2,17 @@
 
 ## Bug Fixes
 
-`artif_spo()` now utilizes `raster::projection()` instead of `raster::crs()` assign a coordinate system to a simple point feature.
+- `artif_spo()` now utilizes `raster::projection()` instead of `raster::crs()` assign a coordinate system to a simple point feature.
 projection(): 
+
+
+# stppSim 1.2.1
+
+Details of changes made (27/03/2022):
+
+1. Added a `vignette` to guide the use of the package
+2. Added a new function, i.e. `compare_boundary.R`
+3. Added a parameter `interactive` to `make_grids.R` function. 
+4. Added new data set,  `birmingham_boundary.rda` inside `inst/extdata`
+5. Modified the `DESCRIPTION` file: R (>= 4.0.0) to R (>= 4.1.0)